Fix input for groups validation in new user form
The hidden input used for form validation was not actually hidden, so it
was shown as a strange input field (it was either empty or it showed the
string representation of the groups array). Although just hiding it will
still keep the validation working when the input is hidden the browser
may not show any validation error in the UI, so it would not be possible
to create the user for "unknown" reasons.

Moreover, the validation API does not provide a way to bind the message
to a different element, and even if the message could be shown using a
Vue tooltip it would not follow the same style as the other native
messages in the rest of the elements.

The Multiselect component internally uses an input element, so now that
one is used for the validation instead of an additional "hidden" input.
That internal element is always empty (the selected groups are shown in
an auxiliary div), so the "required" attribute needs to be explicitly
set and removed on the input based on the currently selected groups.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>